+++ source/extensions/omni.isaac.lab/omni/isaac/lab/envs/direct_rl_env_cfg.py
@@ -41,6 +41,14 @@ class DirectRLEnvCfg:
     """
 
     # general settings
+    seed: int | None = None
+    """The seed for the random number generator. Defaults to None, in which case the seed is not set.
+
+    Note:
+      The seed is set at the beginning of the environment initialization. This ensures that the environment
+      creation is deterministic and behaves similarly across different runs.
+    """
+
     decimation: int = MISSING
     """Number of control action updates @ sim dt per policy dt.
 
@@ -118,4 +126,16 @@ class DirectRLEnvCfg:
     """The noise model applied to the actions provided to the environment. Default is None, which means no noise is added.
 
     Please refer to the :class:`omni.isaac.lab.utils.noise.NoiseModel` class for more details.
+    """
+
+    rerender_on_reset: bool = False
+    """Whether a render step is performed again after at least one environment has been reset.
+    Defaults to False, which means no render step will be performed after reset.
+
+    * When this is False, data collected from sensors after performing reset will be stale and will not reflect the
+      latest states in simulation caused by the reset.
+    * When this is True, an extra render step will be performed to update the sensor data
+      to reflect the latest states from the reset. This comes at a cost of performance as an additional render
+      step will be performed after each time an environment is reset.
+
     """