+++ source/extensions/omni.isaac.lab/omni/isaac/lab/sensors/ray_caster/ray_caster_camera_cfg.py
@@ -5,10 +5,12 @@
 
 """Configuration for the ray-cast camera sensor."""
 
+from dataclasses import MISSING
 from typing import Literal
 
 from omni.isaac.lab.utils import configclass
 
+from .patterns import PinholeCameraPatternCfg
 from .ray_caster_camera import RayCasterCamera
 from .ray_caster_cfg import RayCasterCfg
 
@@ -52,6 +54,9 @@ class RayCasterCameraCfg(RayCasterCfg):
     - ``"none``: No clipping is applied. Values will be returned as ``inf`` for ``distance_to_camera`` and ``nan``
       for ``distance_to_image_plane`` data type.
     """
+
+    pattern_cfg: PinholeCameraPatternCfg = MISSING
+    """The pattern that defines the local ray starting positions and directions in a pinhole camera pattern."""
 
     def __post_init__(self):
         # for cameras, this quantity should be False always.